Authorization server

+++ src/app/api/category/category.controller.ts
@@ -1,10 +1,14 @@
 import dbConnect from "@/database/connection"
 import Category from "@/database/models/category.schema"
+import authMiddleware from "../../../../middleware/auth.middleware"
+import { NextRequest } from "next/server"
 
 export async function createCategory(req:Request){
     try {
-        // const response = authMiddleware(req as NextRequest)
-        // if(response) return response
+        const response = await authMiddleware(req as NextRequest)
+        if(response.status ===401){
+            return response
+        }
         await dbConnect()
         const {name,description} =  await req.json()
         // already exist or not 
@@ -57,9 +61,14 @@ export async function getCategories(){
   }
 }
 
-export async function deleteCategory(id:string){
+export async function deleteCategory(req: Request, id:string){
     try{
         await dbConnect()
+        const response = await authMiddleware(req as NextRequest)
+        if(response.status ===401){
+            return response
+        }
+        
         const deleted = await Category.findByIdAndDelete(id)
         if (!deleted){
             return Response.json({
